Refactor feature dependency ordering into new class
Moved feature ordering logic from ExtensionManager to a new internal DependencyOrdering class. The new implementation centralizes dependency, priority, and ordering constraint handling, improving code organization and maintainability.
